refacotr: separate tool_calling tasks
