+++ src/rai_bench/rai_bench/tool_calling_agent_bench/tasks/manipulation.py
@@ -28,14 +28,11 @@ from rai_bench.tool_calling_agent_bench.interfaces import (
 )
 from rai_bench.tool_calling_agent_bench.mocked_tools import (
     MockGetObjectPositionsTool,
-    MockGetROS2ImageTool,
     MockGetROS2TopicsNamesAndTypesTool,
     MockMoveToPointTool,
-    MockReceiveROS2MessageTool,
 )
 
 loggers_type = logging.Logger
-
 
 PROACTIVE_ROS2_EXPERT_SYSTEM_PROMPT = """You are a ROS 2 expert helping a user with their ROS 2 questions. You have access to various tools that allow you to query the ROS 2 system.
                 Be proactive and use the tools to answer questions.
@@ -46,615 +43,6 @@ class TaskParametrizationError(Exception):
     """Exception raised when the task parameters are not valid."""
 
     pass
-
-
-class GetROS2TopicsTask(ROS2ToolCallingAgentTask):
-    complexity = "easy"
-
-    def __init__(self, logger: loggers_type | None = None) -> None:
-        super().__init__(logger=logger)
-        self.expected_tools = [
-            MockGetROS2TopicsNamesAndTypesTool(
-                mock_topics_names_and_types=[
-                    "topic: /attached_collision_object\ntype: moveit_msgs/msg/AttachedCollisionObject\n",
-                    "topic: /camera_image_color\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /camera_image_depth\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /clock\ntype: rosgraph_msgs/msg/Clock\n",
-                    "topic: /collision_object\ntype: moveit_msgs/msg/CollisionObject\n",
-                    "topic: /color_camera_info\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /color_camera_info5\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /color_image5\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /depth_camera_info5\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /depth_image5\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /display_contacts\ntype: visualization_msgs/msg/MarkerArray\n",
-                    "topic: /display_planned_path\ntype: moveit_msgs/msg/DisplayTrajectory\n",
-                    "topic: /execute_trajectory/_action/feedback\ntype: moveit_msgs/action/ExecuteTrajectory_FeedbackMessage\n",
-                    "topic: /execute_trajectory/_action/status\ntype: action_msgs/msg/GoalStatusArray\n",
-                    "topic: /joint_states\ntype: sensor_msgs/msg/JointState\n",
-                    "topic: /monitored_planning_scene\ntype: moveit_msgs/msg/PlanningScene\n",
-                    "topic: /motion_plan_request\ntype: moveit_msgs/msg/MotionPlanRequest\n",
-                    "topic: /move_action/_action/feedback\ntype: moveit_msgs/action/MoveGroup_FeedbackMessage\n",
-                    "topic: /move_action/_action/status\ntype: action_msgs/msg/GoalStatusArray\n",
-                    "topic: /panda_arm_controller/follow_joint_trajectory/_action/feedback\ntype: control_msgs/action/FollowJointTrajectory_FeedbackMessage\n",
-                    "topic: /panda_arm_controller/follow_joint_trajectory/_action/status\ntype: action_msgs/msg/GoalStatusArray\n",
-                    "topic: /panda_hand_controller/gripper_cmd/_action/feedback\ntype: control_msgs/action/GripperCommand_FeedbackMessage\n",
-                    "topic: /panda_hand_controller/gripper_cmd/_action/status\ntype: action_msgs/msg/GoalStatusArray\n",
-                    "topic: /parameter_events\ntype: rcl_interfaces/msg/ParameterEvent\n",
-                    "topic: /planning_scene\ntype: moveit_msgs/msg/PlanningScene\n",
-                    "topic: /planning_scene_world\ntype: moveit_msgs/msg/PlanningSceneWorld\n",
-                    "topic: /pointcloud\ntype: sensor_msgs/msg/PointCloud2\n",
-                    "topic: /robot_description\ntype: std_msgs/msg/String\n",
-                    "topic: /robot_description_semantic\ntype: std_msgs/msg/String\n",
-                    "topic: /rosout\ntype: rcl_interfaces/msg/Log\n",
-                    "topic: /tf\ntype: tf2_msgs/msg/TFMessage\n",
-                    "topic: /tf_static\ntype: tf2_msgs/msg/TFMessage\n",
-                    "topic: /trajectory_execution_event\ntype: std_msgs/msg/String\n",
-                ]
-            )
-        ]
-
-    def get_system_prompt(self) -> str:
-        return PROACTIVE_ROS2_EXPERT_SYSTEM_PROMPT
-
-    def get_prompt(self) -> str:
-        return "Get the names and types of all ROS2 topics"
-
-    def verify_tool_calls(self, response: dict[str, Any]):
-        """It is expected that the agent will request only the tool that retrieves the ROS2 topics names and types
-
-        Parameters
-        ----------
-        response : dict[str, Any]
-            The response from the agent
-        """
-        messages = response["messages"]
-        ai_messages: Sequence[AIMessage] = [
-            message for message in messages if isinstance(message, AIMessage)
-        ]
-
-        if not ai_messages:
-            self.log_error(msg="No AI messages found in the response.")
-
-        self._is_ai_message_requesting_get_ros2_topics_and_types(ai_messages[0])
-
-        total_tool_calls = sum(len(message.tool_calls) for message in ai_messages)
-        if total_tool_calls != 1:
-            self.log_error(
-                msg=f"Total number of tool calls across all AI messages should be 1, but got {total_tool_calls}."
-            )
-
-        if not self.result.errors:
-            self.result.success = True
-
-
-class GetROS2TopicsTask2(ROS2ToolCallingAgentTask):
-    complexity = "easy"
-
-    def __init__(self, logger: loggers_type | None = None) -> None:
-        super().__init__(logger=logger)
-        self.expected_tools = [
-            MockGetROS2TopicsNamesAndTypesTool(
-                mock_topics_names_and_types=[
-                    "topic: /attached_collision_object\ntype: moveit_msgs/msg/AttachedCollisionObject\n",
-                    "topic: /camera_image_color\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /camera_image_depth\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /clock\ntype: rosgraph_msgs/msg/Clock\n",
-                ]
-            )
-        ]
-
-    def get_system_prompt(self) -> str:
-        return PROACTIVE_ROS2_EXPERT_SYSTEM_PROMPT
-
-    def get_prompt(self) -> str:
-        return "What is in the ROS2 network?"
-
-    def verify_tool_calls(self, response: dict[str, Any]):
-        """It is expected that the agent will request only the tool that retrieves the ROS2 topics names and types
-
-        Parameters
-        ----------
-        response : dict[str, Any]
-            The response from the agent
-        """
-        messages = response["messages"]
-        ai_messages: Sequence[AIMessage] = [
-            message for message in messages if isinstance(message, AIMessage)
-        ]
-
-        if not ai_messages:
-            self.log_error(msg="No AI messages found in the response.")
-
-        self._is_ai_message_requesting_get_ros2_topics_and_types(ai_messages[0])
-
-        total_tool_calls = sum(len(message.tool_calls) for message in ai_messages)
-        if total_tool_calls != 1:
-            self.log_error(
-                msg=f"Total number of tool calls across all AI messages should be 1, but got {total_tool_calls}."
-            )
-
-        if not self.result.errors:
-            self.result.success = True
-
-
-class GetROS2RGBCameraTask(ROS2ToolCallingAgentTask):
-    complexity = "easy"
-
-    def __init__(self, logger: loggers_type | None = None) -> None:
-        super().__init__(logger=logger)
-        self.expected_tools: List[BaseTool] = [
-            MockGetROS2TopicsNamesAndTypesTool(
-                mock_topics_names_and_types=[
-                    "topic: /attached_collision_object\ntype: moveit_msgs/msg/AttachedCollisionObject\n",
-                    "topic: /camera_image_color\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /camera_image_depth\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /clock\ntype: rosgraph_msgs/msg/Clock\n",
-                    "topic: /collision_object\ntype: moveit_msgs/msg/CollisionObject\n",
-                    "topic: /color_camera_info\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /color_camera_info5\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /depth_camera_info5\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /depth_image5\ntype: sensor_msgs/msg/Image\n",
-                ]
-            ),
-            MockGetROS2ImageTool(expected_topics=["/camera_image_color"]),
-        ]
-
-    def get_system_prompt(self) -> str:
-        return PROACTIVE_ROS2_EXPERT_SYSTEM_PROMPT
-
-    def get_prompt(self) -> str:
-        return "Get the RGB image from the camera."
-
-    def verify_tool_calls(self, response: dict[str, Any]):
-        """It is expected that the agent will request:
-        1. The tool that retrieves the ROS2 topics names and types to recognize the RGB image topic
-        2. The tool that retrieves the RGB image from the /camera_image_color topic
-
-        Parameters
-        ----------
-        response : dict[str, Any]
-            The response from the agent
-        """
-        messages = response["messages"]
-        ai_messages: Sequence[AIMessage] = [
-            message for message in messages if isinstance(message, AIMessage)
-        ]
-
-        if len(ai_messages) < 3:
-            self.log_error(
-                msg=f"Expected at least 3 AI messages, but got {len(ai_messages)}."
-            )
-
-        if ai_messages:
-            if not self._is_ai_message_requesting_get_ros2_topics_and_types(
-                ai_messages[0]
-            ):
-                self.log_error(
-                    msg="First AI message did not request ROS2 topics and types correctly."
-                )
-        if len(ai_messages) > 1:
-            if self._check_tool_calls_num_in_ai_message(ai_messages[1], expected_num=1):
-                self._check_tool_call(
-                    tool_call=ai_messages[1].tool_calls[0],
-                    expected_name="get_ros2_image",
-                    expected_args={"topic": "/camera_image_color"},
-                    expected_optional_args={"timeout_sec": None},
-                )
-        if not self.result.errors:
-            self.result.success = True
-
-
-class GetROS2DepthCameraTask(ROS2ToolCallingAgentTask):
-    complexity = "easy"
-
-    def __init__(self, logger: loggers_type | None = None) -> None:
-        super().__init__(logger=logger)
-        self.expected_tools: List[BaseTool] = [
-            MockGetROS2TopicsNamesAndTypesTool(
-                mock_topics_names_and_types=[
-                    "topic: /attached_collision_object\ntype: moveit_msgs/msg/AttachedCollisionObject\n",
-                    "topic: /camera_image_color\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /camera_image_depth\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /clock\ntype: rosgraph_msgs/msg/Clock\n",
-                    "topic: /collision_object\ntype: moveit_msgs/msg/CollisionObject\n",
-                    "topic: /color_camera_info\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /color_camera_info5\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /depth_camera_info5\ntype: sensor_msgs/msg/CameraInfo\n",
-                ]
-            ),
-            MockGetROS2ImageTool(expected_topics=["/camera_image_depth"]),
-        ]
-
-    def get_system_prompt(self) -> str:
-        return PROACTIVE_ROS2_EXPERT_SYSTEM_PROMPT
-
-    def get_prompt(self) -> str:
-        return "Get the depth image from the camera."
-
-    def verify_tool_calls(self, response: dict[str, Any]):
-        """It is expected that the agent will request:
-        1. The tool that retrieves the ROS2 topic names and types to identify the depth image topic.
-        2. The tool that retrieves the RGB image from the /camera_image_depth topic
-
-        Parameters
-        ----------
-        response : dict[str, Any]
-            The response from the agent
-        """
-        messages = response["messages"]
-        ai_messages: Sequence[AIMessage] = [
-            message for message in messages if isinstance(message, AIMessage)
-        ]
-
-        if len(ai_messages) < 3:
-            self.log_error(
-                msg=f"Expected at least 3 AI messages, but got {len(ai_messages)}."
-            )
-
-        if ai_messages:
-            if not self._is_ai_message_requesting_get_ros2_topics_and_types(
-                ai_messages[0]
-            ):
-                self.log_error(
-                    msg="First AI message did not request ROS2 topics and types correctly."
-                )
-
-        if len(ai_messages) > 1:
-            if self._check_tool_calls_num_in_ai_message(ai_messages[1], expected_num=1):
-                self._check_tool_call(
-                    tool_call=ai_messages[1].tool_calls[0],
-                    expected_name="get_ros2_image",
-                    expected_args={"topic": "/camera_image_depth"},
-                    expected_optional_args={"timeout_sec": None},
-                )
-        if not self.result.errors:
-            self.result.success = True
-
-
-class GetAllROS2RGBCamerasTask(ROS2ToolCallingAgentTask):
-    complexity = "easy"
-
-    def __init__(self, logger: loggers_type | None = None) -> None:
-        super().__init__(logger=logger)
-        self.expected_tools: List[BaseTool] = [
-            MockGetROS2TopicsNamesAndTypesTool(
-                mock_topics_names_and_types=[
-                    "topic: /attached_collision_object\ntype: moveit_msgs/msg/AttachedCollisionObject\n",
-                    "topic: /camera_image_color\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /camera_image_depth\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /clock\ntype: rosgraph_msgs/msg/Clock\n",
-                    "topic: /collision_object\ntype: moveit_msgs/msg/CollisionObject\n",
-                    "topic: /color_camera_info\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /color_camera_info5\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /color_image5\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /depth_camera_info5\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /depth_image5\ntype: sensor_msgs/msg/Image\n",
-                ]
-            ),
-            MockGetROS2ImageTool(
-                expected_topics=["/camera_image_color", "/color_image5"]
-            ),
-        ]
-
-    def get_prompt(self) -> str:
-        return "Get RGB images from all of the available cameras."
-
-    def get_system_prompt(self) -> str:
-        return PROACTIVE_ROS2_EXPERT_SYSTEM_PROMPT
-
-    def verify_tool_calls(self, response: dict[str, Any]):
-        """It is expected that the agent will request:
-        1. The tool that retrieves the ROS2 topic names and types to identify the topics with RGB images.
-        2. The tool that retrieves the RGB images - from the /camera_image_color and from the /color_image5 topic
-
-        Parameters
-        ----------
-        response : dict[str, Any]
-            The response from the agent
-        """
-        messages = response["messages"]
-        ai_messages: Sequence[AIMessage] = [
-            message for message in messages if isinstance(message, AIMessage)
-        ]
-
-        if len(ai_messages) < 3:
-            self.log_error(
-                msg=f"Expected at least 3 AI messages, but got {len(ai_messages)}."
-            )
-        if ai_messages:
-            if not self._is_ai_message_requesting_get_ros2_topics_and_types(
-                ai_messages[0]
-            ):
-                self.log_error(
-                    msg="First AI message did not request ROS2 topics and types correctly."
-                )
-
-        if len(ai_messages) > 1:
-            expected_tool_calls: list[dict[str, Any]] = [
-                {
-                    "name": "get_ros2_image",
-                    "args": {"topic": "/camera_image_color"},
-                    "optional_args": {"timeout_sec": None},
-                },
-                {
-                    "name": "get_ros2_image",
-                    "args": {"topic": "/color_image5"},
-                    "optional_args": {"timeout_sec": None},
-                },
-            ]
-
-            self._check_multiple_tool_calls(
-                message=ai_messages[1], expected_tool_calls=expected_tool_calls
-            )
-        if not self.result.errors:
-            self.result.success = True
-
-
-class GetAllROS2DepthCamerasTask(ROS2ToolCallingAgentTask):
-    complexity = "easy"
-
-    def __init__(self, logger: loggers_type | None = None) -> None:
-        super().__init__(logger=logger)
-        self.expected_tools: List[BaseTool] = [
-            MockGetROS2TopicsNamesAndTypesTool(
-                mock_topics_names_and_types=[
-                    "topic: /camera_image_color\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /camera_image_depth\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /clock\ntype: rosgraph_msgs/msg/Clock\n",
-                    "topic: /collision_object\ntype: moveit_msgs/msg/CollisionObject\n",
-                    "topic: /color_camera_info\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /color_camera_info5\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /color_image5\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /depth_camera_info5\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /depth_image5\ntype: sensor_msgs/msg/Image\n",
-                ]
-            ),
-            MockGetROS2ImageTool(
-                expected_topics=["/camera_image_depth", "/depth_image5"]
-            ),
-        ]
-
-    def get_prompt(self) -> str:
-        return "Get depth images from all of the available cameras."
-
-    def get_system_prompt(self) -> str:
-        return PROACTIVE_ROS2_EXPERT_SYSTEM_PROMPT
-
-    def verify_tool_calls(self, response: dict[str, Any]):
-        """It is expected that the agent will request:
-        1. The tool that retrieves the ROS2 topic names and types to identify the topics with depth images.
-        2. The tool that retrieves the depth images - from the /camera_image_depth and from the /depth_image5 topic
-
-        Parameters
-        ----------
-        response : dict[str, Any]
-            The response from the agent
-        """
-        messages = response["messages"]
-        ai_messages: Sequence[AIMessage] = [
-            message for message in messages if isinstance(message, AIMessage)
-        ]
-
-        if len(ai_messages) < 3:
-            self.log_error(
-                msg=f"Expected at least 3 AI messages, but got {len(ai_messages)}."
-            )
-        if ai_messages:
-            if not self._is_ai_message_requesting_get_ros2_topics_and_types(
-                ai_messages[0]
-            ):
-                self.log_error(
-                    msg="First AI message did not request ROS2 topics and types correctly."
-                )
-
-        if len(ai_messages) > 1:
-            expected_tool_calls: list[dict[str, Any]] = [
-                {
-                    "name": "get_ros2_image",
-                    "args": {"topic": "/camera_image_depth"},
-                    "optional_args": {"timeout_sec": None},
-                },
-                {
-                    "name": "get_ros2_image",
-                    "args": {"topic": "/depth_image5"},
-                    "optional_args": {"timeout_sec": None},
-                },
-            ]
-
-            self._check_multiple_tool_calls(
-                message=ai_messages[1], expected_tool_calls=expected_tool_calls
-            )
-        if not self.result.errors:
-            self.result.success = True
-
-
-class GetROS2MessageTask(ROS2ToolCallingAgentTask):
-    complexity = "easy"
-
-    def __init__(self, logger: loggers_type | None = None) -> None:
-        super().__init__(logger=logger)
-        self.expected_tools: List[BaseTool] = [
-            MockGetROS2TopicsNamesAndTypesTool(
-                mock_topics_names_and_types=[
-                    "topic: /attached_collision_object\ntype: moveit_msgs/msg/AttachedCollisionObject\n",
-                    "topic: /camera_image_color\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /camera_image_depth\ntype: sensor_msgs/msg/Image\n",
-                    "topic: /clock\ntype: rosgraph_msgs/msg/Clock\n",
-                    "topic: /collision_object\ntype: moveit_msgs/msg/CollisionObject\n",
-                    "topic: /color_camera_info\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /color_camera_info5\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /depth_camera_info5\ntype: sensor_msgs/msg/CameraInfo\n",
-                    "topic: /depth_image5\ntype: sensor_msgs/msg/Image\n",
-                ]
-            ),
-            MockReceiveROS2MessageTool(expected_topics=["/camera_image_color"]),
-        ]
-
-    def get_system_prompt(self) -> str:
-        return PROACTIVE_ROS2_EXPERT_SYSTEM_PROMPT
-
-    def get_prompt(self) -> str:
-        return "Get RGB image."
-
-    def verify_tool_calls(self, response: dict[str, Any]):
-        """It is expected that the agent will request:
-        1. The tool that retrieves the ROS2 topics names and types to recognize the RGB image topic
-        2. The tool that retrieves the RGB image from the /camera_image_color topic
-
-        Parameters
-        ----------
-        response : dict[str, Any]
-            The response from the agent
-        """
-        messages = response["messages"]
-        ai_messages: Sequence[AIMessage] = [
-            message for message in messages if isinstance(message, AIMessage)
-        ]
-
-        if len(ai_messages) < 3:
-            self.log_error(
-                msg=f"Expected at least 3 AI messages, but got {len(ai_messages)}."
-            )
-        if ai_messages:
-            if not self._is_ai_message_requesting_get_ros2_topics_and_types(
-                ai_messages[0]
-            ):
-                self.log_error(
-                    msg="First AI message did not request ROS2 topics and types correctly."
-                )
-        if len(ai_messages) > 1:
-            if self._check_tool_calls_num_in_ai_message(ai_messages[1], expected_num=1):
-                self._check_tool_call(
-                    tool_call=ai_messages[1].tool_calls[0],
-                    expected_name="receive_ros2_message",
-                    expected_args={"topic": "/camera_image_color"},
-                )
-        if not self.result.errors:
-            self.result.success = True
-
-
-class GetRobotDescriptionTask(ROS2ToolCallingAgentTask):
-    complexity = "easy"
-
-    def __init__(self, logger: loggers_type | None = None) -> None:
-        super().__init__(logger=logger)
-        self.expected_tools: List[BaseTool] = [
-            MockGetROS2TopicsNamesAndTypesTool(
-                mock_topics_names_and_types=[
-                    "topic: /pointcloud\ntype: sensor_msgs/msg/PointCloud2\n",
-                    "topic: /robot_description\ntype: std_msgs/msg/String\n",
-                    "topic: /rosout\ntype: rcl_interfaces/msg/Log\n",
-                    "topic: /tf\ntype: tf2_msgs/msg/TFMessage\n",
-                    "topic: /tf_static\ntype: tf2_msgs/msg/TFMessage\n",
-                    "topic: /trajectory_execution_event\ntype: std_msgs/msg/String\n",
-                ]
-            ),
-            MockReceiveROS2MessageTool(expected_topics=["/robot_description"]),
-        ]
-
-    def get_system_prompt(self) -> str:
-        return PROACTIVE_ROS2_EXPERT_SYSTEM_PROMPT
-
-    def get_prompt(self) -> str:
-        return "Give me description of the robot."
-
-    def verify_tool_calls(self, response: dict[str, Any]):
-        """It is expected that the agent will request:
-        1. The tool that retrieves the ROS2 topics names and types to recognize the topic with the robot description
-        2. The tool that retrieves the message from the /robot_description topic
-
-        Parameters
-        ----------
-        response : dict[str, Any]
-            The response from the agent
-        """
-        messages = response["messages"]
-        ai_messages: Sequence[AIMessage] = [
-            message for message in messages if isinstance(message, AIMessage)
-        ]
-
-        if len(ai_messages) < 3:
-            self.log_error(
-                msg=f"Expected at least 3 AI messages, but got {len(ai_messages)}."
-            )
-        if ai_messages:
-            if not self._is_ai_message_requesting_get_ros2_topics_and_types(
-                ai_messages[0]
-            ):
-                self.log_error(
-                    msg="First AI message did not request ROS2 topics and types correctly."
-                )
-        if len(ai_messages) > 1:
-            if self._check_tool_calls_num_in_ai_message(ai_messages[1], expected_num=1):
-                self._check_tool_call(
-                    tool_call=ai_messages[1].tool_calls[0],
-                    expected_name="receive_ros2_message",
-                    expected_args={"topic": "/robot_description"},
-                )
-        if not self.result.errors:
-            self.result.success = True
-
-
-class GetPointcloudTask(ROS2ToolCallingAgentTask):
-    complexity = "easy"
-
-    def __init__(self, logger: loggers_type | None = None) -> None:
-        super().__init__(logger=logger)
-        self.expected_tools: List[BaseTool] = [
-            MockGetROS2TopicsNamesAndTypesTool(
-                mock_topics_names_and_types=[
-                    "topic: /pointcloud\ntype: sensor_msgs/msg/PointCloud2\n",
-                    "topic: /robot_description\ntype: std_msgs/msg/String\n",
-                    "topic: /rosout\ntype: rcl_interfaces/msg/Log\n",
-                    "topic: /tf\ntype: tf2_msgs/msg/TFMessage\n",
-                    "topic: /tf_static\ntype: tf2_msgs/msg/TFMessage\n",
-                    "topic: /trajectory_execution_event\ntype: std_msgs/msg/String\n",
-                ]
-            ),
-            MockReceiveROS2MessageTool(expected_topics=["/pointcloud"]),
-        ]
-
-    def get_system_prompt(self) -> str:
-        return PROACTIVE_ROS2_EXPERT_SYSTEM_PROMPT
-
-    def get_prompt(self) -> str:
-        return "Get the pointcloud."
-
-    def verify_tool_calls(self, response: dict[str, Any]):
-        """It is expected that the agent will request:
-        1. The tool that retrieves the ROS2 topics names and types to recognize the topic with the pointcloud
-        2. The tool that retrieves the message from the /pointcloud topic
-
-        Parameters
-        ----------
-        response : dict[str, Any]
-            The response from the agent
-        """
-        messages = response["messages"]
-        ai_messages: Sequence[AIMessage] = [
-            message for message in messages if isinstance(message, AIMessage)
-        ]
-
-        if len(ai_messages) < 3:
-            self.log_error(
-                msg=f"Expected at least 3 AI messages, but got {len(ai_messages)}."
-            )
-        if ai_messages:
-            if not self._is_ai_message_requesting_get_ros2_topics_and_types(
-                ai_messages[0]
-            ):
-                self.log_error(
-                    msg="First AI message did not request ROS2 topics and types correctly."
-                )
-        if len(ai_messages) > 1:
-            if self._check_tool_calls_num_in_ai_message(ai_messages[1], expected_num=1):
-                self._check_tool_call(
-                    tool_call=ai_messages[1].tool_calls[0],
-                    expected_name="receive_ros2_message",
-                    expected_args={"topic": "/pointcloud"},
-                )
-        if not self.result.errors:
-            self.result.success = True
 
 
 class MoveToPointTask(ROS2ToolCallingAgentTask):
